Move java.time._ -> plotly.element.LocalDateTime conversions onto the companion, and split out potentially-unsafe conversions so that they require an explicit import.
