+++ core/js/src/main/scala/plotly/element/PlotlyJavaTimeConversions.scala
@@ -1,0 +1,5 @@
+package plotly.element
+
+trait PlotlyJavaTimeConversions {
+  // Empty since the java.time classes are not available in ScalaJS
+}
